requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pendulum==3.0.0
psycopg2==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.34
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.4
